# api/swagger/swagger.yaml
definitions:
...
  render.Error:
    properties:
      details: {}
      error:
        type: string
    type: object
  tag.createRequest:
    properties:
      created_by:
        description: Creator user ID
        type: string
      name:
        description: Tag name
        type: string
    required:
    - name
    type: object
  tag.createResponse:
    properties:
      id:
        description: New tag ID
        type: string
    type: object
  tag.tagResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /admin/tags:
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag payload
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.createResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "409":
          description: Tag already exists
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Create tag
      tags:
      - Tags
  /admin/tags/{id}:
    delete:
      description: Delete a tag by ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid tag ID
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Tag not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Delete tag
      tags:
      - Tags
...
  /tags:
    get:
      description: Returns all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.tagResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: List tags
      tags:
      - Tags
  /tags/{id}:
    get:
      description: Retrieve a tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.tagResponse'
        "400":
          description: Invalid tag ID
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Tag not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Get tag by ID
      tags:
      - Tags
...
swagger: "2.0"